Add teams for kubernetes-sigs/kubebuilder-declarative-pattern
I started the pattern of sig-api-machinery having its own dir
for teams, and moved teams out of kubernetes-sigs/org.yaml that
were related to the server-sdk subproject while I was here. I
also updated their descriptions.

johnsonj is not yet an org member, so I have commented them out
with a ref to the comment asking that they join