Retire from my role as Release Associate
Hey! After a few iterations in sig-release and after a few months helping
as an Associate, I am stepping back.

I recently spoke with a couple of teammates in the sig release and I value
the title I am holding, and I am sure other people from the community
want to contribute as an Associate. Now I have a better idea about how to
contribute and I don't need a title for that.

Signed-off-by: Gianluca Arbezzano <gianarb92@gmail.com>