cmd/merge: use strict unmarshalling for merging configs
Strict unmarshalling allows us to fail when we see an unrecognised
key. This helps prevent subtle issues like `repo` being written instead
of `repos`. Previously, the unmarshalling would have ignored this. This
is quite serious because now this means that the desired state of a team
is that there is no repo under it when there clearly is a repo mentioned.
Hence, peribolos would remove this repo.

Strict unmarshalling prevents us from doing that.

Plus, this uses the yaml packages maintained by the Kubernetes community,
and this is also used in test-infra.

Signed-off-by: Madhav Jivrajani <madhav.jiv@gmail.com>