update sig-cloud-provider owners and leads
this change brings the sig cloud-provider leads information up to date.
Andrew Sy Kim and Nick Turner have become emeritus leads, and as such we
are removing them from the day-to-day duties related to the leads
positions. similarly we are also putting Bridget Kromhout, Walter
Fender, and Michael McCune as the current contacts.